Disable audio offload
When effects are enabled, the usecase for music playback should
always be deep-buffer-playback since we don't have HW effects
enabled. For some reason, the switch from compress-offload-playback
to deep-buffer-playback doesn't happen immediately, causing a
continuous invalidation of the audio track that results in a
momentarily "skipping" playback. Disable audio offload until the
cause of the issue is found.